Add files via upload
Bare bones of a ping pong game vs "AI" using your keyboard Up and down arrow. The game features a reset after every point score. color-coded goals and simple layout and design. I wanted to make the ball speed change based on time or amount of times it has gone back and forth, but scratch as it is - will just make the game more complicated, and glitchy than it needs to be, and as I said, this is a bare minimum of what a pong game is.